requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
pydantic>=2.7
faiss-cpu>=1.8.0
numpy>=1.26
sentence-transformers>=3.0.1
openai>=1.40
python-dotenv>=1.0
